+++ form_utils/widgets.py
@@ -33,25 +33,21 @@ class ImageWidget(forms.FileInput):
         if template is not None:
             self.template = template
         super(ImageWidget, self).__init__(attrs)
-    
+
     def render(self, name, value, attrs=None):
-        from PIL import Image
         input_html = super(forms.FileInput, self).render(name, value, attrs)
-        file_name = str(value)
-        file_path = os.path.join(settings.MEDIA_ROOT, file_name)
-        try: # is image
-            Image.open(file_path)
-            image_html = thumbnail(str(value))
+        if hasattr(value, 'width') and hasattr(value, 'height'):
+            image_html = thumbnail(value.name)
             output = self.template % {'input': input_html,
                                       'image': image_html}
-        except IOError: # not image
+        else:
             output = input_html
         return mark_safe(output)
 
 class ClearableFileInput(forms.MultiWidget):
     default_file_widget_class = forms.FileInput
     template = '%(input)s Clear: %(checkbox)s'
-    
+
     def __init__(self, file_widget=None,
                  attrs=None, template=None):
         if template is not None:
@@ -67,7 +63,7 @@ class ClearableFileInput(forms.MultiWidget):
         else:
             self.value = value
         return super(ClearableFileInput, self).render(name, value, attrs)
-        
+
     def decompress(self, value):
         # the clear checkbox is never initially checked
         return [value, None]
@@ -79,14 +75,14 @@ class ClearableFileInput(forms.MultiWidget):
         return rendered_widgets[0]
 
 root = lambda path: posixpath.join(FORM_UTILS_MEDIA_URL, path)
-    
+
 class AutoResizeTextarea(forms.Textarea):
     """
     A Textarea widget that automatically resizes to accomodate its contents.
-    
+
     """
     class Media:
-        
+
         js = (JQUERY_URL,
               root('form_utils/js/jquery.autogrow.js'),
               root('form_utils/js/autoresize.js'))